% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{get_quantile}
\alias{get_quantile}
\title{The function to calculate Quantiles based on the samples from posterior distribution.}
\usage{
get_quantile(xi, tau, beta, X_pred, test_knots, prob = 0.5)
}
\arguments{
\item{xi}{A value.}

\item{tau}{A value}

\item{beta}{A value}

\item{X_pred}{A value.}

\item{test_knots}{A value.}

\item{prob}{A value.}
}
\value{
A matrix.
}
\description{
The function to calculate Quantiles based on the samples from posterior distribution.
}
\examples{
print("hello world")
}
