% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{get_updates}
\alias{get_updates}
\title{The main MCMC sampler}
\usage{
get_updates(
  xi,
  beta,
  tau,
  CHOL1,
  h_loc,
  X,
  y,
  nmcmc = 2500,
  burnin = 2500,
  thining = 5,
  test_knots,
  verbose = T
)
}
\arguments{
\item{xi}{initial value of xi.}

\item{beta}{initial value of beta.}

\item{tau}{initial value of tau.}

\item{CHOL1}{a matrix.}

\item{h_loc}{the value of h_loc}

\item{X}{design matrix.}

\item{y}{response variable.}

\item{nmcmc}{the number of iterations after burnin.}

\item{burnin}{the number of burnin iterations.}

\item{thining}{the number of thining after the whole sampling is completed.}

\item{test_knots}{the value of test_knots.}

\item{verbose}{TRUE/FALSE. print the log or not.}
}
\value{
Samples from the posterior distribution.
}
\description{
The main MCMC sampler
}
\examples{
print("hello world")
}
