% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticGP-package.R
\docType{package}
\name{logisticGP-package}
\alias{logisticGP}
\alias{logisticGP-package}
\title{logisticGP: This is the title}
\description{
This is the description.
}
\examples{
# example code
print("hello world")
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
