% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{Matern_cov}
\alias{Matern_cov}
\title{The covariance matrix of Matern kernel with general smoothness.}
\usage{
Matern_cov(test_knots, nu = 2.5, sigma = 1, theta = 0.5)
}
\arguments{
\item{test_knots}{A numeric vector.}

\item{nu}{A number.}

\item{sigma}{A number.}

\item{theta}{A number.}
}
\value{
A covariance matrix.
}
\description{
The covariance matrix of Matern kernel with general smoothness.
}
\examples{
Matern_cov(seq(0,1,length.out = 2),2.5,1,0.5)
}
