% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{PostFun_loc}
\alias{PostFun_loc}
\title{A function for inference about posterior distribution.}
\usage{
PostFun_loc(XI_mat, y_grid, X_new, BETA_mat, TAU_mat, location, test_knots)
}
\arguments{
\item{XI_mat}{A matrix.}

\item{y_grid}{value.}

\item{X_new}{value.}

\item{BETA_mat}{value.}

\item{TAU_mat}{value.}

\item{location}{value.}

\item{test_knots}{value.}
}
\value{
A matrix.
}
\description{
A function for inference about posterior distribution.
}
\examples{
print("hello world")
}
