% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{Sigma_kron}
\alias{Sigma_kron}
\title{The Kronecker product of two matrices}
\usage{
Sigma_kron(tau = 1, rho = 0.5, K, p)
}
\arguments{
\item{tau}{A number.}

\item{rho}{A number.}

\item{K}{A integer.}

\item{p}{A integer.}
}
\value{
A matrix.
}
\description{
The Kronecker product of two matrices
}
\examples{
Sigma_kron(1,0.5,4,4)
}
