% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAIC.R
\name{waic.eval}
\alias{waic.eval}
\title{The function to calculate WAIC.}
\usage{
waic.eval(XI_mat, y, X, BETA_mat, TAU_mat, test_knots)
}
\arguments{
\item{XI_mat}{A matrix.}

\item{y}{value.}

\item{X}{value.}

\item{BETA_mat}{A matrix.}

\item{TAU_mat}{A matrix.}

\item{test_knots}{A numeric vector.}
}
\value{
A list.
}
\description{
The function to calculate WAIC.
}
\examples{
print("hello world")
}
